* fix some string to number conversion + added float32 * added battery percentage to devices that have battery

+++ switch.go
@@ -44,7 +44,7 @@ type Switch struct {
 }
 
 type SwitchState struct {
-	Battery float64
+	Battery float32
 	Click   ClickType
 }
